change basetime to double everywhere

+++ uhd_wrapper/include/usrp.hpp
@@ -28,7 +28,7 @@ class Usrp : public UsrpInterface {
     void setTimeToZeroNextPps();
     uint64_t getCurrentSystemTime();
     double getCurrentFpgaTime();
-    void execute(const float baseTime);
+    void execute(const double baseTime);
     std::vector<MimoSignal> collect();
 
     double getMasterClockRate() const { return masterClockRate_; }
@@ -67,8 +67,8 @@ class Usrp : public UsrpInterface {
     void setRxSamplingRate(const double samplingRate, size_t rxAntennaIdx);
 
     void setRfConfigForRxAntenna(const RfConfig& conf, size_t rxAntennaIdx);
-    void transmit(const float baseTime, std::exception_ptr& exceptionPtr);
-    void receive(const float baseTime, std::vector<MimoSignal>& buffers,
+    void transmit(const double baseTime, std::exception_ptr& exceptionPtr);
+    void receive(const double baseTime, std::vector<MimoSignal>& buffers,
                  std::exception_ptr& exceptionPtr);
     void setTimeToZeroNextPpsThreadFunction();
     void processRxStreamingConfig(const RxStreamingConfig& config,
