+++ uhd_wrapper/include/usrp_interface.hpp
@@ -13,7 +13,7 @@ class UsrpInterface {
     virtual void setTimeToZeroNextPps() = 0;
     virtual uint64_t getCurrentSystemTime() = 0;
     virtual double getCurrentFpgaTime() = 0;
-    virtual void execute(const float baseTime) = 0;
+    virtual void execute(const double baseTime) = 0;
     virtual std::vector<MimoSignal> collect() = 0;
     virtual void reset() = 0;
     virtual void resetSyncSources() = 0;
